+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## vNext
+
+- Fix starting local Tailscale
+
 ## 0.6.5
 
 - Delay Funnel's DNS configuration test (new secretfile sub-option)